un-nest RecordConverters
