+++ streams/src/main/java/org/apache/kafka/streams/state/internals/RecordConverter.java
@@ -22,43 +22,4 @@ import java.nio.ByteBuffer;
 
 public interface RecordConverter {
     ConsumerRecord<byte[], byte[]> convert(final ConsumerRecord<byte[], byte[]> record);
-
-    final class RecordConverters {
-        private static final RecordConverter IDENTITY_INSTANCE = record -> record;
-
-        @SuppressWarnings("deprecation")
-        private static final RecordConverter RAW_TO_TIMESTAMED_INSTANCE = record -> {
-            final byte[] rawValue = record.value();
-            final long timestamp = record.timestamp();
-            return new ConsumerRecord<>(
-                record.topic(),
-                record.partition(),
-                record.offset(),
-                timestamp,
-                record.timestampType(),
-                record.checksum(),
-                record.serializedKeySize(),
-                record.serializedValueSize(),
-                record.key(),
-                ByteBuffer
-                    .allocate(8 + rawValue.length)
-                    .putLong(timestamp)
-                    .put(rawValue)
-                    .array(),
-                record.headers(),
-                record.leaderEpoch()
-            );
-        };
-
-        // privatize the constructor so the class cannot be instantiated (only used for its static members)
-        private RecordConverters() {}
-
-        public static RecordConverter rawValueToTimestampedValue() {
-            return RAW_TO_TIMESTAMED_INSTANCE;
-        }
-
-        public static RecordConverter identity() {
-            return IDENTITY_INSTANCE;
-        }
-    }
 }
